Moved SetupPanelController to a higher level.
