+++ src/main/java/ui/controllers/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Contains controllers for the UI
+ */
+package ui.controllers;
