--- src/main/java/ui/debug_panel/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Contains panels for interfacing directly with RS232 components
- */
-package ui.debug_panel;
