--- src/main/java/ui/debug_panel/setup_panel/SetupPanel.java
@@ -1,9 +1,0 @@
-package ui.debug_panel.setup_panel;
-
-import javafx.scene.layout.GridPane;
-
-/**
- * Wrapper for the Setup Panel component
- */
-public class SetupPanel extends GridPane {
-}
