--- src/main/java/ui/debug_panel/setup_panel/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Contains the setup panel component
- */
-package ui.debug_panel.setup_panel;
