--- src/main/resources/debug_panel/SetupPanel.fxml
@@ -1,71 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.layout.GridPane?>
-<?import javafx.scene.text.Font?>
-<?import javafx.scene.text.Text?>
-
-<?import javafx.scene.control.ComboBox?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.layout.HBox?>
-
-<?import javafx.scene.control.TextField?>
-<fx:root alignment="center" type="javafx.scene.layout.GridPane"
-         xmlns:fx="http://javafx.com/fxml/1"
-         fx:controller="ui.debug_panel.setup_panel.SetupPanelController"
-         xmlns="http://javafx.com/javafx/8.0.112">
-    <padding>
-        <Insets bottom="10" left="25" right="25" top="25" />
-    </padding>
-
-    <GridPane alignment="center">
-        <Text text="Connection">
-            <font>
-                <Font size="18.0" />
-            </font>
-            <GridPane.columnIndex>0</GridPane.columnIndex>
-            <GridPane.rowIndex>0</GridPane.rowIndex>
-        </Text>
-        <Text text="Devices">
-            <GridPane.columnIndex>0</GridPane.columnIndex>
-            <GridPane.rowIndex>1</GridPane.rowIndex>
-        </Text>
-        <ComboBox fx:id="availablePortsDropdown">
-            <GridPane.columnIndex>1</GridPane.columnIndex>
-            <GridPane.rowIndex>1</GridPane.rowIndex>
-        </ComboBox>
-
-        <Text text="Custom Address">
-            <GridPane.columnIndex>
-                0
-            </GridPane.columnIndex>
-            <GridPane.rowIndex>
-                2
-            </GridPane.rowIndex>
-        </Text>
-        <TextField fx:id="customAddressField">
-            <GridPane.columnIndex>
-                1
-            </GridPane.columnIndex>
-            <GridPane.rowIndex>
-                2
-            </GridPane.rowIndex>
-        </TextField>
-
-        <Text text="Status">
-            <GridPane.columnIndex>0</GridPane.columnIndex>
-            <GridPane.rowIndex>3</GridPane.rowIndex>
-        </Text>
-        <Text text="Not connected" fx:id="connectionStatus">
-            <GridPane.columnIndex>1</GridPane.columnIndex>
-            <GridPane.rowIndex>3</GridPane.rowIndex>
-        </Text>
-
-        <HBox alignment="BASELINE_RIGHT">
-            <Button fx:id="updateButton" text="Update" onAction="#update" />
-            <Button fx:id="connectButton" text="Connect"/>
-            <GridPane.columnIndex>1</GridPane.columnIndex>
-            <GridPane.rowIndex>4</GridPane.rowIndex>
-        </HBox>
-    </GridPane>
-</fx:root>
