+++ src/test/java/unit/ui/controllers/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Contains unit tests for {@link ui.debug_panel}
  */
-package unit.ui.debug_panel;
+package unit.ui.controllers;
