+++ src/test/java/unit/ui/controllers/setup_panel_controller/SetupPanelControllerTestCase.java
@@ -1,14 +1,14 @@
-package unit.ui.debug_panel.setup_panel;
+package unit.ui.controllers.setup_panel_controller;
 
 import javafx.scene.control.Button;
 import javafx.scene.control.ComboBox;
 import org.junit.Before;
-import unit.ui.debug_panel.DebugPanelTestCase;
+import unit.ui.UserInterfaceTestCase;
 
 /**
- * Created by mkononen on 30/01/17.
+ * Base class for testing {@link ui.controllers.SetupPanelController}
  */
-public abstract class SetupPanelTestCase extends DebugPanelTestCase {
+public abstract class SetupPanelControllerTestCase extends UserInterfaceTestCase {
     protected Button updateButton;
     protected ComboBox<String> availablePortsDropdown;
