+++ src/test/java/unit/ui/controllers/setup_panel_controller/UpdateButton.java
@@ -1,4 +1,4 @@
-package unit.ui.debug_panel.setup_panel;
+package unit.ui.controllers.setup_panel_controller;
 
 import javafx.collections.ObservableList;
 import org.junit.Test;
@@ -11,16 +11,13 @@ import static org.junit.Assert.assertEquals;
  * Tests that clicking the update button successfully updates the list of
  * available ports
  */
-public final class UpdateButton extends SetupPanelTestCase {
+public final class UpdateButton extends SetupPanelControllerTestCase {
     private static final int expected_index = 0;
 
     @Test
     public void testUpdate(){
         List testData = applicationContext.getBean("testData", List.class);
-
-
         clickOn(updateButton);
-
         ObservableList<String> ports = availablePortsDropdown.getItems();
 
         assertEquals(testData.size(), ports.size());
