+++ src/test/java/unit/ui/controllers/setup_panel_controller/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Contains unit tests for {@link ui.controllers.SetupPanelController}
+ */
+package unit.ui.controllers.setup_panel_controller;
