--- src/test/java/unit/ui/debug_panel/DebugPanelTestCase.java
@@ -1,8 +1,0 @@
-package unit.ui.debug_panel;
-
-import unit.ui.UserInterfaceTestCase;
-
-/**
- * Base class for unit tests of {@link ui.debug_panel}
- */
-public abstract class DebugPanelTestCase extends UserInterfaceTestCase {}
