--- src/test/java/unit/ui/debug_panel/setup_panel/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Contains unit tests for {@link ui.debug_panel.setup_panel}
- */
-package unit.ui.debug_panel.setup_panel;
